Do not use nested classes
